cache components loaded from the scope
